requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
cryptography
bitarray
mmh3
pytest
python-dotenv==1.0.0
